Move RecordModuleVersions to metricsbp
