+++ metricsbp/modules.go
@@ -1,4 +1,4 @@
-package prometheusbpint
+package metricsbp
 
 import (
 	"runtime/debug"
@@ -6,13 +6,20 @@ import (
 
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
+
+	"github.com/reddit/baseplate.go/internal/prometheusbpint"
 )
 
-var goModules = promauto.With(GlobalRegistry).NewGaugeVec(prometheus.GaugeOpts{
+var goModules = promauto.With(prometheusbpint.GlobalRegistry).NewGaugeVec(prometheus.GaugeOpts{
 	Name: "baseplate_go_modules",
 	Help: "Export the version information for included Go modules, and whether the module is the 'main' module or a 'dependency'.  Always 1",
 }, []string{"go_module", "module_role", "module_replaced", "module_version"})
 
+// RecordModuleVersions records the modules linked into this binary in the
+// baseplate_go_modules prometheus metric.
+//
+// Users should not need to call this directly, as it is called by baseplate.New.
+// This should generally not be called more than once, and it is not safe to call concurrently.
 func RecordModuleVersions(info *debug.BuildInfo) {
 	record := func(role string, mod *debug.Module) {
 		goModules.With(prometheus.Labels{
